code-examples/backend/group-project-sp23-group10/weather/requirements.txt
Flask==2.2.3
requests
python-dotenv
Flask-Session